pyproject.toml
[project]
name = "litestar-auth-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Yuri Fontella", email = "yurifc4@gmail.com"}
]
requires-python = ">=3.13,<4.0"
dependencies = [
    "litestar[standard] (>=2.17.0,<3.0.0)",
    "litestar-asyncpg (>=0.5.0,<1.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "dnspython (>=2.8.0,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
]

[tool.poetry]
packages = [{ include = "src" }]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.12.12"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.2.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
